Models: Added engine and tyre axis planes.
I want the engines and the tyres to spin around their axis.
However How do I know what their axis are meant to be, without looking it
up and hardcoding it.
So I added a plane perpendicular and centered to the axis of rotation.
Now when I load the object I can not display items labeled as axis and
instead calculate the centre ponit and axis for which to rotate on.