Shadres: Set the background colour to dark blue.
We do this in the deferred (lighting pass) shader, as if you do it in c++
during the geometry pass, the blue ends up in all of your outputs (vertex
postions, normals, and colours).

We do this by looking for normals that are 0,0,0 which is impossible
unless that location hasn't been written to in the frameBuffer, which only
happens if there's nothing there to render.