Rendering Pipeline: Added another stage to my rendering pipeline.
We first do the geometry pass which writes positions, normals and colours
to the geometry pass frame buffer.

Then we have the lighting pass, which calculates the lighting for each
pixel in the geometry pass frame buffer. This gets written to the lighting
pass frame buffer.

Finally we draw all the pixels from the lighting pass frame buffer to the
screen.