Render Pipeline: Add support for a stencil test to not run lighting calculations on the background.
This doesn't seem to have any performance impact on my laptop, probably
because the stencil test is defined to run after the fragment shader.

It can however run before as an optimisation if the fragment shader
doesn't change the depth buffer, so commiting this as an optimisation for
other graphics cards.